==> all_in_one_1/4_embedding(stella).ipynb <==
# NOTE
model `stella_en_1.5B_v5` is one of leader model in huggingface's MTEB(Massive Text Embedding Benchmark) [https://huggingface.co/spaces/mteb/leaderboard]

**input file**: `V:\20240920\all_in_one_1\3_add_emotion.csv`

# embedding

## Loading Data 

from sentence_transformers import SentenceTransformer
import pandas as pd
import numpy as np

print("Loading data...")
df = pd.read_csv(r'V:\20240920\all_in_one_1\3_add_emotion.csv')

# check if exist NaN in column 'letter'. if so, SentenceTransformer report error
print("\nData statistics:")
print(f"Total row number of data: {len(df)}")
print("\n---Check if exist None value in column 'letter':")
print(f"---Number of None values: {df['letter'].isnull().sum()}")
   
# Creating embeddings
print("Creating embeddings...")
texts = df['letter'].tolist()

print("\n !!! The input data for SentenceTransformer() must be list !!!")
print(f"      Verified the type of variable texts is :  {type(texts)}")

df.head(1)

## do embeddings

print("Beginning  embeddings...")

from sentence_transformers import SentenceTransformer

# using model all-mpnet-base-v2
# model = SentenceTransformer(r'V:\huggingface\model\sentence-transformers\all-mpnet-base-v2')
# embeddings = model.encode(texts, show_progress_bar=True, max_length=2048)

# using model stella_en_1.5B_v5
# ！The default dimension is 1024, if you need other dimensions, please clone the model and modify `modules.json` to replace `2_Dense_1024` with another dimension, e.g. `2_Dense_256` or `2_Dense_8192` !
# model = SentenceTransformer(r"V:\huggingface\model\MTEB\stella_en_1.5B_v5", trust_remote_code=True).cuda()
model = SentenceTransformer(r"V:\huggingface\model\MTEB\stella_en_1.5B_v5", trust_remote_code=True)
embeddings = model.encode(texts, show_progress_bar=True, max_length=2048)

print(f"The embeddings shape is: {embeddings.shape}")

## save embeddings

# save embeddings 
print("\nSaving embeddings...")
# np.save(r"V:\20240920\way2\2_embeddings_stella_en_1.5B_v5.npy", embeddings)
# vnp.save(r"V:\20240920\way2\2_embeddings_all-mpnet-base-v2.npy", embeddings)
np.save(r"V:\20240920\all_in_one_1\4_embeddings_stella.npy", embeddings)
